+++ NEWS.md
@@ -1,6 +1,7 @@
 # simstandard 0.2.0
 
-* Added the `fixed2free` function which takes a `lavaan` model with fixed parameters and returns a model in which all parameters are free.
+* Added the `fixed2free` function, which takes a `lavaan` syntax model with fixed parameters and returns a `lavaan` syntax model in which all parameters are free.
+* Added the `model_complete` function, which takes a `lavaan` syntax model with standardized loadings, structure coefficients, and covariances, and returns a `lavaan` syntax model with all standardized coefficients, including standardized variances.
 * Added a `NEWS.md` file to track changes to the package.
 
 # simstandard 0.1.0